Pass PYTORCH_ROCM_ARCH to FA build
The Dockerfile layer that builds flash attention uses the
PYTORCH_ROCM_ARCH argument.  But since we don't have
ARG PYTORCH_ROCM_ARCH under the FROM for that layer, the argument
isn't actually in scope and therefore isn't set.  This causes
flash attention to try to build with the default list of targets
(a much longer list than the two architectures we use by default
in our PYTORCH_ROCM_ARCH, and the default list is currently failing
for the ck_tile branch of FA).